requirements.txt
pyrogram
requests
moviepy
Pillow
